# docs/swagger.yaml
definitions:
...
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: string
    type: object
...
  key.req:
    properties:
      description:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/token:
    get:
      consumes:
      - application/json
      description: Get all API keys for the authenticated user with their status and
        metadata
      produces:
      - application/json
      responses:
        "200":
          description: API key list
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: List API keys
      tags:
      - keys
    post:
      consumes:
      - application/json
      description: Create a new API key for programmatic access
      parameters:
      - description: API key creation request (description is optional)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/key.req'
      produces:
      - application/json
      responses:
        "200":
          description: API key created successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Create API key
      tags:
      - keys
...
swagger: "2.0"